build: reformat code

+++ src/daemon/priv.c
@@ -39,15 +39,15 @@
 #include <netinet/if_ether.h>
 
 #ifdef HAVE_LINUX_CAPABILITIES
-#include <sys/capability.h>
-#include <sys/prctl.h>
+#  include <sys/capability.h>
+#  include <sys/prctl.h>
 #endif
 
 #if defined HOST_OS_FREEBSD || defined HOST_OS_OSX || defined HOST_OS_DRAGONFLY
-# include <net/if_dl.h>
+#  include <net/if_dl.h>
 #endif
 #if defined HOST_OS_SOLARIS
-# include <sys/sockio.h>
+#  include <sys/sockio.h>
 #endif
 
 /* Use resolv.h */
@@ -69,11 +69,11 @@
 
 /* Bionic has res_init() but it's not in any header */
 #if defined HAVE_RES_INIT && defined __BIONIC__
-int res_init (void);
+int res_init(void);
 #endif
 
 #ifdef ENABLE_PRIVSEP
-static int monitored = -1;		/* Child */
+static int monitored = -1; /* Child */
 #endif
 
 /* Proxies */
@@ -111,15 +111,12 @@ priv_gethostname()
 	must_write(PRIV_UNPRIVILEGED, &cmd, sizeof(enum priv_cmd));
 	priv_wait();
 	must_read(PRIV_UNPRIVILEGED, &len, sizeof(int));
-	if (len < 0 || len > 255)
-		fatalx("privsep", "too large value requested");
-	if ((buf = (char*)realloc(buf, len+1)) == NULL)
-		fatal("privsep", NULL);
+	if (len < 0 || len > 255) fatalx("privsep", "too large value requested");
+	if ((buf = (char *)realloc(buf, len + 1)) == NULL) fatal("privsep", NULL);
 	must_read(PRIV_UNPRIVILEGED, buf, len);
 	buf[len] = '\0';
 	return buf;
 }
-
 
 int
 priv_iface_init(int index, char *iface)
@@ -187,8 +184,7 @@ priv_snmp_socket(struct sockaddr_un *addr)
 	must_write(PRIV_UNPRIVILEGED, addr, sizeof(struct sockaddr_un));
 	priv_wait();
 	must_read(PRIV_UNPRIVILEGED, &rc, sizeof(int));
-	if (rc < 0)
-		return rc;
+	if (rc < 0) return rc;
 	return receive_fd(PRIV_UNPRIVILEGED);
 }
 
@@ -207,10 +203,8 @@ asroot_ctl_cleanup()
 	int rc = 0;
 
 	must_read(PRIV_PRIVILEGED, &len, sizeof(int));
-	if (len < 0 || len > PATH_MAX)
-		fatalx("privsep", "too large value requested");
-	if ((ctlname = (char*)malloc(len+1)) == NULL)
-		fatal("privsep", NULL);
+	if (len < 0 || len > PATH_MAX) fatalx("privsep", "too large value requested");
+	if ((ctlname = (char *)malloc(len + 1)) == NULL) fatal("privsep", NULL);
 
 	must_read(PRIV_PRIVILEGED, ctlname, len);
 	ctlname[len] = 0;
@@ -226,27 +220,24 @@ static void
 asroot_gethostname()
 {
 	struct utsname un;
-	struct addrinfo hints = {
-		.ai_flags = AI_CANONNAME
-	};
+	struct addrinfo hints = { .ai_flags = AI_CANONNAME };
 	struct addrinfo *res;
 	int len;
-	if (uname(&un) < 0)
-		fatal("privsep", "failed to get system information");
+	if (uname(&un) < 0) fatal("privsep", "failed to get system information");
 	if (getaddrinfo(un.nodename, NULL, &hints, &res) != 0) {
 		log_info("privsep", "unable to get system name");
 #ifdef HAVE_RES_INIT
 		res_init();
 #endif
-                len = strlen(un.nodename);
-                must_write(PRIV_PRIVILEGED, &len, sizeof(int));
-                must_write(PRIV_PRIVILEGED, un.nodename, len);
-        } else {
-                len = strlen(res->ai_canonname);
-                must_write(PRIV_PRIVILEGED, &len, sizeof(int));
-                must_write(PRIV_PRIVILEGED, res->ai_canonname, len);
+		len = strlen(un.nodename);
+		must_write(PRIV_PRIVILEGED, &len, sizeof(int));
+		must_write(PRIV_PRIVILEGED, un.nodename, len);
+	} else {
+		len = strlen(res->ai_canonname);
+		must_write(PRIV_PRIVILEGED, &len, sizeof(int));
+		must_write(PRIV_PRIVILEGED, res->ai_canonname, len);
 		freeaddrinfo(res);
-        }
+	}
 }
 
 static void
@@ -262,7 +253,7 @@ asroot_iface_init()
 	TRACE(LLDPD_PRIV_INTERFACE_INIT(name));
 	rc = asroot_iface_init_os(ifindex, name, &fd);
 	must_write(PRIV_PRIVILEGED, &rc, sizeof(rc));
-	if (rc == 0 && fd >=0) send_fd(PRIV_PRIVILEGED, fd);
+	if (rc == 0 && fd >= 0) send_fd(PRIV_PRIVILEGED, fd);
 	if (fd >= 0) close(fd);
 }
 
@@ -286,19 +277,19 @@ asroot_iface_multicast()
 	must_read(PRIV_PRIVILEGED, LLADDR(dlp), ETHER_ADDR_LEN);
 #elif defined HOST_OS_OPENBSD || defined HOST_OS_NETBSD || defined HOST_OS_SOLARIS
 	struct sockaddr *sap = (struct sockaddr *)&ifr.ifr_addr;
-#if ! defined HOST_OS_SOLARIS
+#  if !defined HOST_OS_SOLARIS
 	sap->sa_len = sizeof(struct sockaddr);
-#endif
+#  endif
 	sap->sa_family = AF_UNSPEC;
 	must_read(PRIV_PRIVILEGED, sap->sa_data, ETHER_ADDR_LEN);
 #else
-#error Unsupported OS
+#  error Unsupported OS
 #endif
 
 	must_read(PRIV_PRIVILEGED, &add, sizeof(int));
 	if (((sock = socket(AF_INET, SOCK_DGRAM, 0)) == -1) ||
-	    ((ioctl(sock, (add)?SIOCADDMULTI:SIOCDELMULTI,
-		    &ifr) < 0) && (errno != EADDRINUSE)))
+	    ((ioctl(sock, (add) ? SIOCADDMULTI : SIOCDELMULTI, &ifr) < 0) &&
+		(errno != EADDRINUSE)))
 		rc = errno;
 
 	if (sock != -1) close(sock);
@@ -314,10 +305,8 @@ asroot_iface_description()
 	must_read(PRIV_PRIVILEGED, &name, sizeof(name));
 	name[sizeof(name) - 1] = '\0';
 	must_read(PRIV_PRIVILEGED, &len, sizeof(int));
-	if (len < 0 || len > PATH_MAX)
-		fatalx("privsep", "too large value requested");
-	if ((description = (char*)malloc(len+1)) == NULL)
-		fatal("privsep", NULL);
+	if (len < 0 || len > PATH_MAX) fatalx("privsep", "too large value requested");
+	if ((description = (char *)malloc(len + 1)) == NULL) fatal("privsep", NULL);
 
 	must_read(PRIV_PRIVILEGED, description, len);
 	description[len] = 0;
@@ -355,28 +344,28 @@ asroot_snmp_socket()
 		must_read(PRIV_PRIVILEGED, &bogus, sizeof(struct sockaddr_un));
 	if (addr->sun_family != AF_UNIX)
 		fatal("privsep", "someone is trying to trick me");
-	addr->sun_path[sizeof(addr->sun_path)-1] = '\0';
+	addr->sun_path[sizeof(addr->sun_path) - 1] = '\0';
 
 	if ((sock = socket(PF_UNIX, SOCK_STREAM, 0)) < 0) {
 		log_warn("privsep", "cannot open socket");
 		must_write(PRIV_PRIVILEGED, &sock, sizeof(int));
 		return;
 	}
-        if ((rc = connect(sock, (struct sockaddr *) addr,
-		    sizeof(struct sockaddr_un))) != 0) {
-		log_info("privsep", "cannot connect to %s: %s",
-                          addr->sun_path, strerror(errno));
+	if ((rc = connect(sock, (struct sockaddr *)addr, sizeof(struct sockaddr_un))) !=
+	    0) {
+		log_info("privsep", "cannot connect to %s: %s", addr->sun_path,
+		    strerror(errno));
 		close(sock);
 		rc = -1;
 		must_write(PRIV_PRIVILEGED, &rc, sizeof(int));
 		return;
-        }
+	}
 
 	int flags;
 	if ((flags = fcntl(sock, F_GETFL, NULL)) < 0 ||
 	    fcntl(sock, F_SETFL, flags | O_NONBLOCK) < 0) {
 		log_warn("privsep", "cannot set sock %s to non-block : %s",
-                          addr->sun_path, strerror(errno));
+		    addr->sun_path, strerror(errno));
 
 		close(sock);
 		rc = -1;
@@ -391,23 +380,20 @@ asroot_snmp_socket()
 
 struct dispatch_actions {
 	enum priv_cmd msg;
-	void(*function)(void);
+	void (*function)(void);
 };
 
-static struct dispatch_actions actions[] = {
-	{PRIV_PING, asroot_ping},
-	{PRIV_DELETE_CTL_SOCKET, asroot_ctl_cleanup},
-	{PRIV_GET_HOSTNAME, asroot_gethostname},
+static struct dispatch_actions actions[] = { { PRIV_PING, asroot_ping },
+	{ PRIV_DELETE_CTL_SOCKET, asroot_ctl_cleanup },
+	{ PRIV_GET_HOSTNAME, asroot_gethostname },
 #ifdef HOST_OS_LINUX
-	{PRIV_OPEN, asroot_open},
+	{ PRIV_OPEN, asroot_open },
 #endif
-	{PRIV_IFACE_INIT, asroot_iface_init},
-	{PRIV_IFACE_MULTICAST, asroot_iface_multicast},
-	{PRIV_IFACE_DESCRIPTION, asroot_iface_description},
-	{PRIV_IFACE_PROMISC, asroot_iface_promisc},
-	{PRIV_SNMP_SOCKET, asroot_snmp_socket},
-	{-1, NULL}
-};
+	{ PRIV_IFACE_INIT, asroot_iface_init },
+	{ PRIV_IFACE_MULTICAST, asroot_iface_multicast },
+	{ PRIV_IFACE_DESCRIPTION, asroot_iface_description },
+	{ PRIV_IFACE_PROMISC, asroot_iface_promisc },
+	{ PRIV_SNMP_SOCKET, asroot_snmp_socket }, { -1, NULL } };
 
 /* Main loop, run as root */
 static void
@@ -418,10 +404,10 @@ priv_loop(int privileged, int once)
 
 #ifdef ENABLE_PRIVSEP
 	setproctitle("monitor.");
-#ifdef USE_SECCOMP
+#  ifdef USE_SECCOMP
 	if (priv_seccomp_init(privileged, monitored) != 0)
-	   fatal("privsep", "cannot continue without seccomp setup");
-#endif
+		fatal("privsep", "cannot continue without seccomp setup");
+#  endif
 #endif
 	while (!may_read(PRIV_PRIVILEGED, &cmd, sizeof(enum priv_cmd))) {
 		log_debug("privsep", "received command %d", cmd);
@@ -431,8 +417,7 @@ priv_loop(int privileged, int once)
 				break;
 			}
 		}
-		if (a->function == NULL)
-			fatalx("privsep", "bogus message received");
+		if (a->function == NULL) fatalx("privsep", "bogus message received");
 		if (once) break;
 	}
 }
@@ -441,17 +426,18 @@ priv_loop(int privileged, int once)
  * the other case, it should be called when we wait an action from the
  * privileged side. */
 void
-priv_wait() {
+priv_wait()
+{
 #ifndef ENABLE_PRIVSEP
 	/* We have no remote process on the other side. Let's emulate it. */
 	priv_loop(0, 1);
 #endif
 }
 
-
 #ifdef ENABLE_PRIVSEP
 static void
-priv_exit_rc_status(int rc, int status) {
+priv_exit_rc_status(int rc, int status)
+{
 	switch (rc) {
 	case 0:
 		/* kill child */
@@ -496,8 +482,7 @@ static void
 sig_pass_to_chld(int sig)
 {
 	int oerrno = errno;
-	if (monitored != -1)
-		kill(monitored, sig);
+	if (monitored != -1) kill(monitored, sig);
 	errno = oerrno;
 }
 
@@ -519,7 +504,8 @@ sig_chld(int sig)
 }
 
 /* Create a subdirectory and check if it's here. */
-static int _mkdir(const char *pathname, mode_t mode)
+static int
+_mkdir(const char *pathname, mode_t mode)
 {
 	int save_errno;
 	if (mkdir(pathname, mode) == 0 || errno == EEXIST) {
@@ -540,9 +526,10 @@ static int _mkdir(const char *pathname, mode_t mode)
 }
 
 /* Create a directory recursively. */
-static int mkdir_p(const char *pathname, mode_t mode)
+static int
+mkdir_p(const char *pathname, mode_t mode)
 {
-	char path[PATH_MAX+1];
+	char path[PATH_MAX + 1];
 	char *current;
 
 	if (strlcpy(path, pathname, sizeof(path)) >= sizeof(path)) {
@@ -554,18 +541,16 @@ static int mkdir_p(const char *pathname, mode_t mode)
 	for (current = path + 1; *current; current++) {
 		if (*current != '/') continue;
 		*current = '\0';
-		if (_mkdir(path, mode) != 0)
-			return -1;
+		if (_mkdir(path, mode) != 0) return -1;
 		*current = '/';
 	}
-	if (_mkdir(path, mode) != 0)
-		return -1;
+	if (_mkdir(path, mode) != 0) return -1;
 
 	return 0;
 }
 
 /* Initialization */
-#define LOCALTIME "/etc/localtime"
+#  define LOCALTIME "/etc/localtime"
 static void
 priv_setup_chroot(const char *chrootdir)
 {
@@ -577,12 +562,10 @@ priv_setup_chroot(const char *chrootdir)
 	/* Check if /etc/localtime exists in chroot or outside chroot */
 	char path[1024];
 	int source = -1, destination = -1;
-	if (snprintf(path, sizeof(path),
-		"%s" LOCALTIME, chrootdir) >= sizeof(path))
+	if (snprintf(path, sizeof(path), "%s" LOCALTIME, chrootdir) >= sizeof(path))
 		return;
 	if ((source = open(LOCALTIME, O_RDONLY)) == -1) {
-		if (errno == ENOENT)
-			return;
+		if (errno == ENOENT) return;
 		log_warn("privsep", "cannot read " LOCALTIME);
 		return;
 	}
@@ -591,8 +574,7 @@ priv_setup_chroot(const char *chrootdir)
 	path[strlen(chrootdir) + 4] = '\0';
 	if (mkdir(path, 0755) == -1) {
 		if (errno != EEXIST) {
-			log_warn("privsep", "unable to create %s directory",
-			    path);
+			log_warn("privsep", "unable to create %s directory", path);
 			close(source);
 			return;
 		}
@@ -603,10 +585,9 @@ priv_setup_chroot(const char *chrootdir)
 	char buffer[1024];
 	ssize_t n;
 	mode_t old = umask(S_IWGRP | S_IWOTH);
-	if ((destination = open(path,
-		    O_WRONLY | O_CREAT | O_TRUNC | O_EXCL, 0666)) == -1) {
-		if (errno != EEXIST)
-			log_warn("privsep", "cannot create %s", path);
+	if ((destination = open(path, O_WRONLY | O_CREAT | O_TRUNC | O_EXCL, 0666)) ==
+	    -1) {
+		if (errno != EEXIST) log_warn("privsep", "cannot create %s", path);
 		close(source);
 		umask(old);
 		return;
@@ -645,7 +626,8 @@ static void
 sig_chld(int sig)
 {
 	int status = 0;
-	while (waitpid(-1, &status, WNOHANG) > 0);
+	while (waitpid(-1, &status, WNOHANG) > 0)
+		;
 }
 
 #endif
@@ -657,20 +639,15 @@ priv_drop(uid_t uid, gid_t gid)
 	gidset[0] = gid;
 	log_debug("privsep", "dropping privileges");
 #ifdef HAVE_SETRESGID
-	if (setresgid(gid, gid, gid) == -1)
-		fatal("privsep", "setresgid() failed");
+	if (setresgid(gid, gid, gid) == -1) fatal("privsep", "setresgid() failed");
 #else
-	if (setregid(gid, gid) == -1)
-		fatal("privsep", "setregid() failed");
+	if (setregid(gid, gid) == -1) fatal("privsep", "setregid() failed");
 #endif
-	if (setgroups(1, gidset) == -1)
-		fatal("privsep", "setgroups() failed");
+	if (setgroups(1, gidset) == -1) fatal("privsep", "setgroups() failed");
 #ifdef HAVE_SETRESUID
-	if (setresuid(uid, uid, uid) == -1)
-		fatal("privsep", "setresuid() failed");
+	if (setresuid(uid, uid, uid) == -1) fatal("privsep", "setresuid() failed");
 #else
-	if (setreuid(uid, uid) == -1)
-		fatal("privsep", "setreuid() failed");
+	if (setreuid(uid, uid) == -1) fatal("privsep", "setreuid() failed");
 #endif
 }
 
@@ -683,11 +660,13 @@ priv_caps(uid_t uid, gid_t gid)
 		"cap_dac_override,cap_net_raw,cap_net_admin,cap_setuid,cap_setgid=pe",
 		"cap_dac_override,cap_net_raw,cap_net_admin=pe"
 	};
-	log_debug("privsep", "getting CAP_NET_RAW/ADMIN and CAP_DAC_OVERRIDE privilege");
+	log_debug("privsep",
+	    "getting CAP_NET_RAW/ADMIN and CAP_DAC_OVERRIDE privilege");
 	if (!(caps = cap_from_text(caps_strings[0])))
 		fatal("privsep", "unable to convert caps");
 	if (cap_set_proc(caps) == -1) {
-		log_warn("privsep", "unable to drop privileges, monitor running as root");
+		log_warn("privsep",
+		    "unable to drop privileges, monitor running as root");
 		cap_free(caps);
 		return;
 	}
@@ -729,8 +708,7 @@ priv_init(void)
 
 #ifdef ENABLE_PRIVSEP
 	/* Spawn off monitor */
-	if ((monitored = fork()) < 0)
-		fatal("privsep", "unable to fork monitor");
+	if ((monitored = fork()) < 0) fatal("privsep", "unable to fork monitor");
 	switch (monitored) {
 	case 0:
 		/* We are in the children, drop privileges */
@@ -740,8 +718,7 @@ priv_init(void)
 			priv_setup_chroot(chrootdir);
 			if (chroot(chrootdir) == -1)
 				fatal("privsep", "unable to chroot");
-			if (chdir("/") != 0)
-				fatal("privsep", "unable to chdir");
+			if (chdir("/") != 0) fatal("privsep", "unable to chdir");
 			priv_drop(uid, gid);
 		}
 		close(pair[1]);
@@ -757,31 +734,25 @@ priv_init(void)
 		priv_caps(uid, gid);
 
 		/* Install signal handlers */
-		const struct sigaction pass_to_child = {
-			.sa_handler = sig_pass_to_chld,
-			.sa_flags = SA_RESTART
-		};
+		const struct sigaction pass_to_child = { .sa_handler = sig_pass_to_chld,
+			.sa_flags = SA_RESTART };
 		sigaction(SIGALRM, &pass_to_child, NULL);
 		sigaction(SIGTERM, &pass_to_child, NULL);
-		sigaction(SIGHUP,  &pass_to_child, NULL);
-		sigaction(SIGINT,  &pass_to_child, NULL);
+		sigaction(SIGHUP, &pass_to_child, NULL);
+		sigaction(SIGINT, &pass_to_child, NULL);
 		sigaction(SIGQUIT, &pass_to_child, NULL);
-		const struct sigaction child = {
-			.sa_handler = sig_chld,
-			.sa_flags = SA_RESTART
-		};
+		const struct sigaction child = { .sa_handler = sig_chld,
+			.sa_flags = SA_RESTART };
 		sigaction(SIGCHLD, &child, NULL);
-		sig_chld(0);	/* Reap already dead children */
+		sig_chld(0); /* Reap already dead children */
 		priv_loop(pair[1], 0);
 		exit(0);
 	}
 #else
-	const struct sigaction child = {
-		.sa_handler = sig_chld,
-		.sa_flags = SA_RESTART
-	};
+	const struct sigaction child = { .sa_handler = sig_chld,
+		.sa_flags = SA_RESTART };
 	sigaction(SIGCHLD, &child, NULL);
-	sig_chld(0);	/* Reap already dead children */
+	sig_chld(0); /* Reap already dead children */
 	log_warnx("priv", "no privilege separation available");
 	priv_ping();
 #endif
